snb_bot/requirements.txt
Django==2.1
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
requests==2.19.1
